fix(domain): finalize input validation and number parsing behavior

+++ src/test/java/domain/SumCalculatorTest.java
@@ -15,7 +15,7 @@ class SumCalculatorTest {
     @Test
     @DisplayName("합산 계산 테스트")
     void sumCalculatorTest() {
-        List<Integer> numList = Arrays.asList(1,3,5,7);
+        List<Double> numList = Arrays.asList(1.0,3.0,5.0,7.0);
         assertEquals(16, sumCalculator.sumCalculator(numList));
     }
 }
